+++ src/main/java/roomescape/booking/waiting/WaitingRepository.java
@@ -2,14 +2,13 @@ package roomescape.booking.waiting;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import roomescape.booking.schedule.Schedule;
-import roomescape.member.Member;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface WaitingRepository extends JpaRepository<Waiting, Long> {
 
-    List<Waiting> findAllByMember(Member member);
+    List<Waiting> findAllByMember_Email(String email);
 
     Waiting findFirstByScheduleOrderByCreatedAtAsc(Schedule schedule);
